+++ gui/static/js/components/gridview.js
@@ -1,0 +1,13 @@
+var React = require('react');
+
+export default class GridView extends React.Component {
+    //TODO
+    componentDidMount(){
+        //main();
+    }
+    render() {
+        return(
+            <div id ="component_view" className = "box">GridView</div>
+        )
+    }
+}
