+++ gui/static/js/components/scenarios.js
@@ -1,0 +1,10 @@
+var React = require('react');
+
+export default class Scenarios extends React.Component {
+    //TODO
+    render() {
+        return(
+            <div id ="component_scenarios" className = "box">Scenarios</div>
+        )
+    }
+}
